Improving the closedness of the TyC class.
Before it was possible to create new instances but it required leaving the reify method undefined.
With this method it isn't possible to add new instaces at all.

Technically with extensions you can add overlapping instances but they must necessarily be identical to the instances they overlap because there are not methods available to override.