Move from a cryptohash-sha256 dependency to a SHA dependency.
Although cryptohash-sha256 is designed for fast hash computaiton, the interface doesn't let us access SHA-256 midstates.
The SHA library, although pure Haskell, has an interface that is flexable enough to provide access to SHA-256 midstates.

We will use the SHA library for now, and we can patch cryptohash-sha256 to provide access to midstates in the future.