Refactor computeEvalTCOBounds into eval.c
This ensures that the memory bound are computed correctly directly before evaluation.
This is particularly helpful when 'forceJets' (see next commit) is used which might modify the dag.
The disadvantange is that these bounds are not cached when the same expression is evaluated multiplie times.
However, for production use where only Simplicity programs are evaluated, each program is only ever evaluated once.
For testing purposes the bounds are quickly recomputed so shouldn't be a problem.
If it does become a problem, this code can be refactored latered to allow for caching of these bounds.