Reduce memory use of static analysis tests.
The simplifier was consuming a lot of memory when compiling the 'hashBlock' static analysis test.
The problem was during the instantiation of the Arbitrary (Word256, Word512) instance,
which expanded out the 786 instances of Arbitrary 'Either () ()' and all the (nested) pairs of those instances.

Now we generate Arbitrary integers instead.

The '-O2 -with-rtsopts=-c' build options keep the runtime memory use of running the tests down
by enabling optimizations and turning on the compacting garbage collector.

Futher examination on how to reduce memory use is still warranted, but this will do for now.