Fixups for off-curve points.
Fixups of secp256k1 to make it compatible with our prefered specifications.
(1) Our specification has no infinity flag, and treats the infinity flag as a cache of a z=0 test.
This view holds everywhere except when doubling a point of order 2.
We fix this instance.
We also adjust secp256k1_gej_set_ge and secp256k1_gej_add_zinv_var to ensure that the z-value is set to zero when infinity is returned.

(2) gej_add_ge leaves the rzr value uninitialized when the a input is infinity.
This can occur when ecmult builds a scalar table for very low order points, i.e. odd order points
of order less than 15 or so.
While this uninitialized value is read and then eventually multiplied by 0, this is techincally unspecified behaviour.
Thus we prefer to explicity clear the rzr in this case.

(3) Our specification of GE cannot include points at infinity.
However secp256k1_ge_globalz_set_table_gej may produce GE points at the infinity when ecmult builds a scalar table for very low order points,
i.e. odd order points of order less than 15 or so.
For compatability with our specification, we clear the infinity flag, yeilding the cusp point of (0,0).