Fix decoding error in Boolean jets
The decoding of the boolean case should be at code 1, not code 0.

I've changed the Haskell code so that putPositive crashes when given a
non-positive value, rather than silently putting 1 value.  This probably would
have prevented the error, and indeed the Haskell tests fail with only this
change to putPositive.