Add option to not fail on timing tests
CPU timing is inherently a flaky test.

The '--no-timing' flag will make it so that results of timing tests are ignored,
although the timing results will still be printed to standard out.

The '--timing' flag will include the results of the timing tests.

By deault '--no-timing' will be used unless compiled with -DTIMING_FLAG,
in which case the default will be '--timing'.