Sort in memcmp order.
Previously we'd sort by comparing the least significant byte first.
For deciding if duplicates exist, the natures of the ordering doesn't matter.

However now we want to sort assetID in order to lookup fees in a binary search.
To do that lookup, we perfer to use memcmp order.

memcmp order is still implementation defined, however this order is not
observable through the "hasDuplicate" or binary search lookup interfaces, so it
is still safe to use in consensus.