Assert preconditions for execSimplicity
Currently a 'false' result in execSimplicity is ambigous.
It either means a precondition failed, or it failed to complete its computation
due to lack of memory or some other "transient" error.

But failing a preconditon is not "transient".  Instead we will just assert the
various precondtions hold and require that the caller ensure that the
preconditions are satisfied.

We also add a new 'non-negative budget' precondition that was missing before.