Safer linear combination and scale jets
These safer jets requires the input point to be on-curve.

There is some concern that the libsecp256k1's ecmult implemetion might not be
memory safe when run on off-curve points of very low order.
While I'm fairly certain this doesn't happen, it is probably better to err on
the safe side.