Update return type of Bitcoin time lock jets
Using the type aliases clarifies the output variants of these jets.

All Bitcoin jets (with primitives or without) should have access to
the same type aliases. It might be useful to define the alises in the
chapter on Bitcoin + Simplicity, similar to how the Elements aliases are
defined in the chapter on Elements + Simplicity.