Fix bindings to scale and linear_combination_1
When the safe_ versions of these functions were introduced, the
primitiveJetNode.inc file was not correctly updated to bind to the new "safe_"
version of the jets.

To remedy this issue we instead exchange the names of the safe_ functions,
removing the safe_ prefix and instead adding an "off_curve_" prefix to the
unsafe versions.

primitiveJetNode.inc isn't touched in this commit since it was already
referencing the names without a prefix.