Fixup empty arrays in test vectors
Arrays cannot be of zero size in C.

Normally I would change them to NULL pointers, however one cannot assign a NULL
value to an array.  I could, and maybe should, change the header so that these
values are pointers and not array, but if I were to do that, I ought to change
all of the types of the test vectors to be pointers, in order to be consistent.
In which case to assign an array to the pointer I'd have to use something like
a compound literal.

That seems like a big pain for just test vectors, so for now we are just going
to put a null character into what should be empty arrays and explicity set their
corresponding size value to 0.