Fix malforemd Confidential(ONE)
Confidential(ONE), and hence its deriviative Asset(ONE) and Amount(ONE),
contains a record structure, which isn't part of the Simplicity type system, and
hence cannot be used in Simplicity types.

This reworks separate definitions for Asset and AssetWith (and similarly for
Amount/AmountWith) such that Asset and AssetWith are proper Simplicity types.