Correctly name identity hashes
What we have been calling Identity Merkle Roots in the C and Haskell codes is
not the IMR, and instead is a hash of the IMR and the Type Merkle Roots of the
input and output types.

To avoid further confusion we appropriately rename these values to be Identity
Hashes instead.