Generate contents of decodePrimitive automatically
The new generation code isn't exactly the same as the orginal code.
There is no optimistic lifting out of a seconde decoding (code2 variable).
Such optimizations could be added back in if wanted, but it is probably best to
keep the code straightforward.

The decodePrimitive is already being tested by the "Serialization.C" tests in the
Haskell test suite via FFI bindings.