Remove unneed UBOUNDED_MAX checks
For the case of COMP, if type_dag[COMP_B(...)].bitSize exceeds UNBOUNDED_MAX,
then the bounded_add and friends functions will already set extraCellsBound[0]
and extraCellsBound[1] to UBOUNDED_MAX.  Therefore the branch doesn't have any
computational effect.  It is also not a case worth optimizing for.

The case of DISCONNECT is similar, however possibly only one of
extraCellsBound[0] or extraCellsBound[1] will end up being set to UBOUNDED_MAX
instead of both.

This is fine, because these values are just intermediate values for
the *cellsBound computation at the end of the function. There we take the
bounded_max of both the extraCellsBound[0] and extraCellsBound[1] functions. So
the result is the same whether or not both extraCellsBound values are maxed out
or only one is.