Make primitive.h functions indirect
The original plan here was to build different Simplicity primitives (e.g.
Elements and Bitcoin), which are selected at link time.

However, we would like to link multiple sets of primitives into Haskell for
testing purposes.  This means we will need different names for the Bitcoin and
Elements to bind them.

To support this we turn the primitive.h functions into callbacks for the various
specific primitive implementations.

Various functions are rearranged to support the removal of primitive.h and make
the pairing of .c file implementing .h header files more consistent.  It should
now be the case that each .c files implement exactly the declairations of one
.h file.